divide logic, add controllers and schemas
