+++ schemas/contactSchema.js
@@ -1,0 +1,9 @@
+const Joi = require('joi');
+
+const contactSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().required(),
+  phone: Joi.string().required(),
+});
+
+module.exports = contactSchema;
